17-Health_Check/requirements.txt
fastapi[all]
uvicorn
gunicorn
pydantic
asyncpg
sqlalchemy
redis==4.3.3
fastapi-cache2[redis]==0.1.8
fastapi-healthcheck==0.2.8
psutil